requirements.txt
python-telegram-bot
pymongo
python-dotenv
colorlog
motor
aiodns
uvloop
httpx
